Crate bmw_test

The BMW Test Utility crate

This crate provides macros that are used in the tests in other crates within the BMW repo.

Macros

free_port

Returns a free port that is not used at the time of the call. It is also guaranteed to not be allocated to another test which calls this macro or the crate::test_info macro, so there are no timing concerns.

test_info

Macro to setup a test directory based on the function name. A free port is also returned. The directory is removed when the returned value goes out of scope unless the preserve value is specifed and set to true. Specifically a crate::TestInfo is returned by this macro.

Structs

Builder

A builder that is used to construct TestInfo implementations. This is typically called through the crate::test_info macro.

Constants

QA_SLEEP

A sleep time designed to allow execution of tests to run to completion. We avoid using this sleep parameter, but it is needed in a few cases.

Traits

TestInfo

This trait defines the data which a test can use. A test can obtain one of these by calling the crate::test_info macro.

Functions

pick_free_port

Pick a free port that does not collide with recently assigned ports. This is used in tests throughout various BMW crates.