Trait bmw_test::TestInfo

pub trait TestInfo {
    // Required methods
    fn port(&self) -> u16;
    fn directory(&self) -> &String;
    fn sync_channel(&self) -> (SyncSender<()>, Receiver<()>);
}

This trait defines the data which a test can use. A test can obtain one of these by calling the crate::test_info macro.

Required Methods

fn port(&self) -> u16

Return a port that can be used by the test.

fn directory(&self) -> &String

Return a directory that can be used by the test. It is automatically deleted when the crate::TestInfo goes out of scope.

fn sync_channel(&self) -> (SyncSender<()>, Receiver<()>)

Return a (std::sync::mpsc::SyncSender<()>, std::sync::mpsc::Receiver<()>) in which the sender will automatically send message after 60 seconds. This allows threads to timeout. This is useful in eventhandler / http / rustlet testing.